backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Validation and Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# AI - Doubao/OpenAI compatible
openai==1.10.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
Pillow==10.2.0
openpyxl==3.1.2

# Optional - for production
gunicorn==21.2.0

pillow==12.2.0
